+++ driven/storage/adapter.go
@@ -183,7 +183,7 @@ func (sa *Adapter) getCachedServiceRegs(serviceIDs []string) ([]model.ServiceReg
 
 	var serviceRegList []model.ServiceReg
 	var err error
-	if !logutils.ContainsString(serviceIDs, "all") {
+	if !utils.Contains(serviceIDs, "all") {
 		serviceRegList = make([]model.ServiceReg, len(serviceIDs))
 		for i, serviceID := range serviceIDs {
 			item, _ := sa.cachedServiceRegs.Load(serviceID)
@@ -2164,6 +2164,27 @@ func (sa *Adapter) DeleteAppOrgRole(id string) error {
 	if deletedCount == 0 {
 		return errors.WrapErrorData(logutils.StatusMissing, model.TypeAppOrgRole, &logutils.FieldArgs{"_id": id}, err)
 	}
+	return nil
+}
+
+//InsertAppOrgRolePermissions inserts permissions to role
+func (sa *Adapter) InsertAppOrgRolePermissions(context TransactionContext, roleID string, permissions []model.Permission) error {
+
+	filter := bson.D{primitive.E{Key: "_id", Value: roleID}}
+	update := bson.D{
+		primitive.E{Key: "$push", Value: bson.D{
+			primitive.E{Key: "permissions", Value: bson.M{"$each": permissions}},
+		}},
+	}
+
+	res, err := sa.db.applicationsOrganizationsRoles.UpdateOne(filter, update, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionFind, model.TypeAppOrgRole, nil, err)
+	}
+	if res.ModifiedCount != 1 {
+		return errors.ErrorAction(logutils.ActionUpdate, model.TypeAppOrgRole, &logutils.FieldArgs{"unexpected modified count": res.ModifiedCount})
+	}
+
 	return nil
 }